backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
google-generativeai
lime
python-multipart
scikit-learn
numpy
